feat: added routing and pages
