+++ package.json
@@ -22,7 +22,8 @@
     "clsx": "2.1.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-hot-toast": "2.4.1"
+    "react-hot-toast": "2.4.1",
+    "react-router-dom": "6.27.0"
   },
   "devDependencies": {
     "@commitlint/cli": "19.5.0",
